Application/server/requirements.txt
requests
pandas
numpy
matplotlib
tensorflow
scikit-learn
fastapi
pydantic
python-dotenv